Made category for NSDate date conversion; cleaned up style; made use of static_cast and null checking macros; added default cases to switch statements; updated date logic to account for missing year/month; moved include files.
