+++ Frameworks/AddressBook/NSDate+AddressBookAdditions.mm
@@ -14,18 +14,25 @@
 //
 //******************************************************************************
 
-#import "ABContactInternal.h"
+#import <NSDate+AddressBookAdditions.h>
 #import "UWP/WindowsApplicationModelContacts.h"
 
-@implementation _ABContact
+@implementation NSDate (AddressBookAdditions)
 
-- (id)initWithContact:(WACContact*)contact {
-    self = [super init];
-    if (self) {
-        self.contact = contact;
-    }
++ (NSDate*)dateWithWACContactDate:(WACContactDate*)date {
+    NSCalendar* calendar = [NSCalendar currentCalendar];
+    NSDateComponents* dateComponents = [[NSDateComponents alloc] init];
 
-    return self;
+    // Grab the numerical values of the day/month/year,
+    // and create an NSDate to match the same day as the
+    // given NSDate. Default month of January and year
+    // of 2004 are per the guidelines provided from
+    // Windows.ApplicationModel.Contacts.ContactDate
+    dateComponents.day = [date.day integerValue];
+    dateComponents.month = date.month ? [date.month integerValue] : 1;
+    dateComponents.year = date.year ? [date.year integerValue] : 2004;
+    NSDate* resultDate = [calendar dateFromComponents:dateComponents];
+    return resultDate;
 }
 
 @end
